account for pending payments and store preimages of paid invoices.

+++ call_functions.go
@@ -17,7 +17,7 @@ func calcCallCosts(c *types.Call) {
 	c.Cost = s.FixedCallCostSatoshis * 1000
 	c.Cost += int(float64(len(c.Payload)) * 10)
 	if c.Satoshis > 100 {
-		c.Cost += c.Satoshis / 200
+		c.Cost += int(float64(c.Satoshis) / 200)
 	}
 }
